Fixed a packet sent by LoginServer
Changed length from byte to uint.

+++ TSO LoginServer/TSO LoginServer/Network/PacketHandlers.cs
@@ -66,7 +66,7 @@ namespace TSO_LoginServer.Network
                 //0x01 = InitLoginNotify
                 PacketStream OutPacket = new PacketStream(0x01, 2);
                 OutPacket.WriteByte(0x01);
-                OutPacket.WriteByte(0x01);
+                OutPacket.WriteUInt16(0x01);
                 Client.Username = AccountName;
                 //This is neccessary to encrypt packets.
                 Client.Password = Account.GetPassword(AccountName);
@@ -78,7 +78,7 @@ namespace TSO_LoginServer.Network
             {
                 PacketStream OutPacket = new PacketStream(0x02, 2);
                 P.WriteByte(0x02);
-                P.WriteByte(0x01);
+                P.WriteUInt16(0x01);
                 Client.Send(P.ToArray());
 
                 Logger.LogInfo("Bad accountname - sent SLoginFailResponse!\r\n");
